Fix: Crouch Bug + Prep: Dash
Fixed a bug where the player kept spamming the crouch start animation, due to the bool becoming false before the transition idle animation - that bool was not necessary and was removed.
Also added Event Handler and bool variable for a future Dash and DashAttack move.
You might get the warning, that these variables are never used. This is not a problem.